+++ cartridges/int_mercadopago/cartridge/client/default/js/checkout/billing.js
@@ -45,12 +45,17 @@ base.methods.updatePaymentInformation = (order) => {
           htmlToAppend += "</div>";
         } else if (selectedPaymentInstrument.paymentMethod === "CHECKOUT_PRO") {
           const mercadoPagoImage = $(".checkout-pro-option")[0].currentSrc;
+          const mercadoPagoText = $(".mp-text-messages").data("mpTextMessages")["field.chopro.billing.message"];
           htmlToAppend += "<img src='" + mercadoPagoImage + "' ";
           htmlToAppend += "height='32' ";
           htmlToAppend += "alt='CHECKOUT_PRO' ";
           htmlToAppend += "title='CHECKOUT_PRO' ";
           htmlToAppend += "style='margin-top: 10px' ";
           htmlToAppend += "/>";
+          htmlToAppend += "<div>";
+          htmlToAppend += "<p style='line-height: 21px;margin-top: 10px'>" + mercadoPagoText;
+          htmlToAppend += "</p>";
+          htmlToAppend += "</div>";
         } else if (selectedPaymentInstrument.paymentMethod === "MERCADO_CREDITO") {
           const mercadoPagoImage = $(".checkout-pro-option")[0].currentSrc;
           const mercadoPagoText = $(".mp-text-messages").data("mpTextMessages")["field.mercadocredito.billing.message"];