+++ e2e/flows/delete_one_saved_card.js
@@ -1,0 +1,5 @@
+export const deleteOneSavedCard = async (page) => {
+  await page.getByLabel('Delete Payment', { exact: true }).first().click();
+  await page.waitForSelector('.modal-dialog', { timeout: 5000 } )
+  await page.getByText('Yes').click();
+};